requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
pymysql==1.1.1
cryptography==44.0.3
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.11.3